Only reset images on hard reset otherwise seed images would be lost before first test run

+++ update.sh
@@ -5,9 +5,7 @@ shopt -s extglob
 rm -rf !(wp-config.php|wp-cypress-config.php|seeds|wp-content|update.sh|.htaccess)
 
 # Ensure uploads work
-rm -rf /var/www/html/wp-content/uploads
-mkdir -p /var/www/html/wp-content/uploads
-chmod -R 777 /var/www/html/wp-content/uploads
+if ${HARD_RESET}; then rm -rf /var/www/html/wp-content/uploads && mkdir -p /var/www/html/wp-content/uploads && chmod -R 777 /var/www/html/wp-content/uploads; fi
 
 if ${HARD_RESET}; then rm wp-config.php; fi
